develop in progress

+++ sql/CreateTableScripts.sql
@@ -9,8 +9,8 @@ create table if not exists products
     product_id      numeric primary key,
     article         varchar,
     product_name    varchar,
-    last_buy_price  numeric,
-    last_sell_price numeric,
+    last_buy_price  double precision,
+    last_sell_price double precision,
     amount          numeric,
     storage_id      numeric,
     constraint FK_storage_id foreign key (storage_id) references storages (storage_id)
